chore(repo): move git check to conformance
